requirements.txt
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.1
Markdown==3.3.6
MarkupSafe==2.1.1
Pillow==9.1.0
PyJWT==2.3.0
requests==2.27.1
SQLAlchemy==1.4.35
sqlparse==0.4.2
dj-static
Unipath
python-decouple
bleach
psycopg2
whitenoise
